requirements.txt
asgiref==3.8.1
black==24.4.2
click==8.1.7
Django==5.0.6
djangorestframework==3.15.1
isort==5.13.2
mypy-extensions==1.0.0
numpy==1.26.4
opencv-python==4.10.0.82
packaging==24.1
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
python-dotenv==1.0.1
sqlparse==0.5.0
